+++ CHANGELOG.md
@@ -10,6 +10,14 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+### [4.0.5] - 2024-11-28
+
+### Fix
+
+- update rich-domain lib to check nullish type: now 'create' return a possibly null value in Result instance.
+
+---
+
 ### [4.0.4] - 2024-09-26
 
 ### Fix