+++ packages/cpf/CHANGELOG.md
@@ -8,6 +8,14 @@ All notable changes to this project will be documented in this file.
 
 ## Released
 
+---
+
+### [0.0.3] - 2024-11-28
+
+### Fix
+
+- update rich-domain lib to check nullish type: now 'create' return a possibly null value in Result instance.
+
 ### [0.0.2] - 2024-05-31
 
 ### Docs